+++ backend/models/BookingLog.go
@@ -1,9 +1,9 @@
 package models
 
 type BookingLog struct {
-	ID        int64
-	BookingID int64
-	UserID    int64
+	ID        int
+	BookingID int
+	UserID    int
 	State     string
 	Details   string
 }
