+++ backend/models/BusinessConfig.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type BusinessConfig struct {
-	ID         int64
-	BusinessID int64
+	ID         int
+	BusinessID int
 	Key        string
 	Value      string
 	CreatedAt  time.Time
