+++ backend/models/Communication.go
@@ -3,9 +3,9 @@ package models
 import "time"
 
 type Communication struct {
-	ID        int64
-	ChannelID int64
-	UserID    int64
+	ID        int
+	ChannelID int
+	UserID    int
 	From      string
 	To        string
 	Content   string
