+++ backend/models/Contact.go
@@ -1,9 +1,9 @@
 package models
 
 type Contact struct {
-	ID              int64
+	ID              int
 	ContactableType string
-	ContactableID   int64
+	ContactableID   int
 	Address         string
 	Phone           string
 	Email           string
